requirements.txt
pandas>=1.3.0
pdfminer.six>=20221105
openpyxl>=3.0.0
streamlit>=1.20.0
pyyaml>=6.0
python-docx>=0.8.11
opencv-python-headless>=4.5.0
Pillow>=9.0.0
pytesseract>=0.3.10
openai>=1.0.0
google-generativeai>=0.3.0
